requirements.txt
fastapi
uvicorn
python-multipart
jinja2
pyyaml